Fix regexp following a keyword
This reflects similar rule to what Acorn and Esprima use in JS parsing world - if you see a keyword, and it's not `this`, it can be followed only by either regular expression or be a syntax error (we don't check the latter).

This might still run into edge cases such as `yield` being an identifier in legacy code, in which case it has to be followed by division, but, as with any contextual keywords, this would require more complexity.

See https://github.com/tdewolff/minify/issues/132 for the impact of this change.